phidata/requirements.txt
phidata 
python-dotenv
yfinance
packaging
duckduckgo-search
fastapi
uvicorn
groq
openai
sqlalchemy
pgvector
psycopg
pypdf
phidata
google-generativeai
duckduckgo-search
streamlit